Test if the database is empty before installing DeskPRO
We need the DeskPRO files in place before we can install the software.
But at the same time, we can't install the software if the database is
not empty. The installer also does this, but in this image checking if
the db is empty beforehand allows us to keep the target directory empty
for the next run if needed (if the database is emptied afterwards)