Split the `docker-compose.yml` file into their variants
The `docker-compose.yml` file was supposed to be used for testing, but
since it had a generic name, it could be confused as "the way to run
Deskpro".

This commit splits the file into two, one using only the fpm variant
and one using apache. The files have `-test-` in their name to suggest
they're used for testing only. Also, each one uses a differently named
volume since otherwise the `deskpro` volume would be the same for the
two "projects" unless some extra options were passed in from the command
line, which isn't ideal.

While this requires a `-f filename.yml` switch on the command line, it
should reduce confusion.