Install runtime libraries manually too
Some automatic dependencies of the dev packages were being removed
during `apt-get remove`. e.g. `libpng16-16`, which is a dependency of
`libpng-dev`. It was installed with `libpng-dev`, but since it was
installed as a dependency, `apt-get remove libpng-dev` will remove the
dependency too.

This commit installs some of these dependencies manually so that when we
remove their dev counterparts the runtimes are still present on the
system.