Add simple retry for gpg
The keyserver used for gosu is somewhat unstable. More frequently than
not the `--recv-keys` command will fail.

To increase the reliability of the building process a bit, a very simple
retry mechanism has been added: retry up to 5 times, and if the command
fails wait a few seconds. If all retries fail, the next command will
fail and the build will finish.